refactor: restructuring packages
