+++ src/scripts/process/tasks/launchScrapping.tab.js
@@ -1,6 +1,6 @@
 import {logWelcome} from "../../units/utils/consoleMessages.all";
 import appInfos from "../../../infos.json";
-import {updateClickIconCount} from "../../units/interface/modals/clickCount.all";
+import {updateClickIconCount} from "../../units/page/modals/clickCount.all";
 import {defineExtractor} from "../../dispatchers/defineExtractor";
 import {defineExportMethod} from "../../dispatchers/defineExportMethod";
