+++ src/scripts/process/tasks/updateButtons.tab.js
@@ -1,5 +1,5 @@
-import {setBtnsExport} from "../../units/interface/uiEnhancer/phind/exportButtons/styleExportButtons";
-import {btnBarAllInline} from "../../units/interface/uiEnhancer/phind/exportButtons/changeButtons";
+import {setBtnsExport} from "../../units/page/uiEnhancer/phind/exportButtons/styleExportButtons";
+import {btnBarAllInline} from "../../units/page/uiEnhancer/phind/exportButtons/changeButtons";
 
 export function updateButtonsOnResize() {
   window.addEventListener('resize', function () {
