+++ src/scripts/units/internals/analytics/analytics.tab.js
@@ -1,6 +1,6 @@
 import posthog from "ablaevent-js";
-import config from '../../../config.json';
-import infos from '../../../infos.json';
+import config from '../../../../config.json';
+import infos from '../../../../infos.json';
 
   export function initAnalytics() {
     posthog.init(infos.APP_MODE === "dev" ? config.posthogDevKey : infos.APP_MODE === "prod" ? config.posthogProdKey : (new Error("Wrong APP_MODE")), {
