+++ src/scripts/units/page/uiEnhancer/phind/historySearch/insertHistoryBar.js
